# unfurl/server/openapi.json
{
  "components": {
    "schemas": {
...
      "CloudMapDocument": {
        "properties": {
          "apiVersion": {
            "enum": [
              "unfurl/v1alpha1",
              "unfurl/v1.0.0"
            ],
            "type": "string"
          },
          "artifacts": {
            "additionalProperties": {
              "$ref": "#/components/schemas/cloudmap_artifact"
            },
            "description": "Tangible object that instantiates services or other artifacts. Artifact ID is either a package URL (see <https://github.com/package-url/purl-spec>) or repository URL with path.",
            "propertyNames": {
              "description": "A PURL identifier or another URL if the artifact is of a type not supported by PURL.",
              "pattern": "^[^\\s]*$"
            },
            "type": "object"
          },
          "instantiations": {
            "additionalProperties": {
              "$ref": "#/components/schemas/cloudmap_instantiation"
            },
            "description": "Build and deployment information for artifacts and services. Keys are URLs.",
            "propertyNames": {
              "description": "URL of the instantiation.",
              "pattern": "^[^\\s]*$"
            },
            "type": "object"
          },
          "kind": {
            "const": "CloudMap"
          },
          "metadata": {
            "$ref": "#/components/schemas/cloudmap_metadata",
            "description": "Human-readable metadata about this CloudMap."
          },
          "repositories": {
            "additionalProperties": {
              "$ref": "#/components/schemas/cloudmap_repository"
            },
            "description": "Git repositories. Keys are URLs that start with git://",
            "propertyNames": {
              "description": "URL of the repository using the git:// URL scheme.",
              "pattern": "^[^\\s]*$"
            },
            "type": "object"
          },
          "services": {
            "additionalProperties": {
              "$ref": "#/components/schemas/cloudmap_service"
            },
            "description": "Instances of services.",
            "propertyNames": {
              "description": "URL of the service.",
              "pattern": "^[^\\s]*$"
            },
            "type": "object"
          },
          "types": {
            "additionalProperties": {
              "$ref": "#/components/schemas/cloudmap_type"
            },
            "description": "Type definitions for artifacts, services, software, and capabilities.",
            "propertyNames": {
              "description": "Fully-qualified type name with namespace (e.g., Compute@unfurl.cloud/onecommons/std).",
              "pattern": "^[^\\s]+$"
            },
            "type": "object"
          }
        },
        "required": [
          "apiVersion",
          "kind"
        ],
        "title": "CloudMap Schema",
        "type": "object"
      },
      "CloudMapDocumentPair": {
        "description": "Two-element array: the queried (and optionally filtered) CloudMap document followed by a CloudMap document of records discovered by following the graph from the filter key (empty dict when ``follow`` is 0 or no key was supplied).",
        "items": {
          "$ref": "#/components/schemas/CloudMapDocument"
        },
        "maxItems": 2,
        "minItems": 2,
        "title": "CloudMap document pair",
        "type": "array"
      },
...
      "HTTPError": {
        "properties": {
          "detail": {
            "type": "object"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "detail": {
            "properties": {
              "<location>": {
                "properties": {
                  "<field_name>": {
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  }
                },
                "type": "object"
              }
            },
            "type": "object"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Unfurl Server API",
    "version": "1.2.1"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/cloudmap": {
      "get": {
        "description": "Return a pair ``[document, follow]`` of CloudMap documents. ``document`` is the raw CloudMap (or a subset filtered by ``kind`` / ``key``). ``follow`` contains records reachable from ``key`` when ``follow`` > 0, otherwise it is ``{}``.",
        "parameters": [
          {
            "in": "query",
            "name": "auth_project",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "description": "Project ID for authorization and cache key scoping",
              "title": "Auth Project"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "repositories",
                    "artifacts",
                    "services",
                    "instantiations",
                    "types"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "description": "Top-level CloudMap section to return; if omitted the full document is returned.",
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "description": "Record key (URL) within the selected ``kind`` section; ignored when ``kind`` is omitted.",
              "title": "Key"
            }
          },
          {
            "in": "query",
            "name": "follow",
            "required": false,
            "schema": {
              "default": 0,
              "description": "If > 0 and ``key`` is supplied, walk the CloudMap graph starting at ``key`` and return the discovered records in the second element of the response pair. Otherwise the second element is an empty dict.",
              "title": "Follow",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CloudMapDocumentPair"
                }
              }
            },
            "description": "Pair: [filtered CloudMap document, followed records]"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "CloudMap document",
        "tags": [
          "Export"
        ]
      },
      "post": {
        "description": "Apply a batch of add / update / delete operations to ``cloudmap.yaml``. Top-level keys split between an envelope (``latest_commit`` / ``cloudmap_path`` / ``username`` / ``private_token`` / ``commit_msg``) and the cloudmap sections (``repositories``, ``artifacts``, ``services``, ``instantiations``, ``types``).\n\nEach section maps record keys to a JSON object that schema-validates as the corresponding cloudmap entity. To delete a record, send the object with ``unfurl.server.deleted: true``.\n\nThe body is validated against ``docs/cloudmap-schema.json`` (a 422 is returned on schema violation). On success the file is committed locally (no push) and the new commit oid is returned.",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloudMapDocument"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchResponse"
                }
              }
            },
            "description": "commit oid produced by the local commit, or null when nothing changed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Modify the CloudMap document",
        "tags": [
          "Export"
        ]
      }
    },
    "/create_ensemble": {
      "post": {
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchEnsembleBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Conflict (repository at wrong revision)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "summary": "Create a new ensemble",
        "tags": [
          "Project"
        ]
      }
    },
    "/create_provider": {
      "post": {
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchEnsembleBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Conflict (repository at wrong revision)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "summary": "Create a cloud provider and its associated ensemble",
        "tags": [
          "Project"
        ]
      }
    },
    "/delete_deployment": {
      "post": {
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchEnvironmentBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Conflict (repository at wrong revision)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "summary": "Delete a deployment",
        "tags": [
          "Project"
        ]
      }
    },
    "/delete_environment": {
      "post": {
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchEnvironmentBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Conflict (repository at wrong revision)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "summary": "Delete a deployment environment",
        "tags": [
          "Project"
        ]
      }
    },
...
    "/export": {
      "get": {
        "description": "Export an ensemble or service template in a JSON format suitable for the frontend. Supports 'deployment', 'blueprint', and 'environments' formats.",
        "parameters": [
          {
            "in": "query",
            "name": "auth_project",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "description": "Project ID for authorization and cache key scoping",
              "title": "Auth Project"
            }
          },
          {
            "in": "query",
            "name": "latest_commit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "description": "Commit hash used to validate the cache entry",
              "title": "Latest Commit"
            }
          },
          {
            "in": "query",
            "name": "branch",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "description": "Git branch name",
              "title": "Branch"
            }
          },
          {
            "in": "query",
            "name": "pretty",
            "required": false,
            "schema": {
              "default": false,
              "description": "Pretty-print the JSON response",
              "title": "Pretty",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "username",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "description": "Git username (alternative to X-Git-Credentials header)",
              "title": "Username"
            }
          },
          {
            "in": "query",
            "name": "visibility",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "public",
                    "private"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "description": "Repository visibility",
              "title": "Visibility"
            }
          },
          {
            "in": "query",
            "name": "queueid",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "description": "If set, the Rust proxy will enqueue the request for async processing via Redis instead of proxying synchronously",
              "title": "Queueid"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "deployment",
              "description": "Export format",
              "enum": [
                "deployment",
                "blueprint",
                "environments"
              ],
              "title": "Format",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "deployment_path",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "description": "Path to the deployment within the project",
              "title": "Deployment Path"
            }
          },
          {
            "in": "query",
            "name": "environment",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "description": "Environment name (used with 'environments' format)",
              "title": "Environment"
            }
          },
          {
            "in": "query",
            "name": "include_all_deployments",
            "required": false,
            "schema": {
              "default": false,
              "description": "Include all deployment exports embedded in the response",
              "title": "Include All Deployments",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "stale",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "const": "ok",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "description": "Return any cache hit without checking if it's out of date.",
              "title": "Stale"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportResponse"
                }
              }
            },
            "description": "GraphQL-style JSON database of TOSCA objects"
          },
          "304": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Not Modified (ETag matched)"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "summary": "Export ensemble as JSON",
        "tags": [
          "Export"
        ]
      }
    },
...
    "/update_ensemble": {
      "post": {
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchEnsembleBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Conflict (repository at wrong revision)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "summary": "Update an existing ensemble",
        "tags": [
          "Project"
        ]
      }
    },
    "/update_environment": {
      "post": {
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchEnvironmentBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Conflict (repository at wrong revision)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "summary": "Update a deployment environment",
        "tags": [
          "Project"
        ]
      }
    },
...
  },
...
}